producers/wiki/requirements.txt
kafka-python==2.0.2
requests==2.32.3
sseclient-py==1.8.0
